Add token_uri to kcs-5
This adds the token_uri read endpoint to kcs-5. The purpose of this is for if a collection does not know what their metadata will be for each token when the collection is created but rather on mint and wants to store their metadata off-chain. The on-chain metadata feature also solves this same problem, but, only if the collection owner wants to store their metadata on chain. I am not willing to debate which method is "better" as it isn't up to me what people store on chain or off chain. There is an upcoming collection that utilizes this feature so it seems to make sense to add it to the KCS-5 standard.